ref: Reorganize layout components
