+++ src/components/layout/ErrorLayout.tsx
@@ -1,18 +1,20 @@
-import { Box, Text, Heading } from "@chakra-ui/react";
+import { Text, Heading } from "@chakra-ui/react";
 import { useRouteError } from "react-router-dom";
 
-export default function ErrorPage() {
+import { DefaultLayout } from "./DefaultLayout";
+
+export function ErrorLayout() {
   // TODO: Type here?
   const error = useRouteError();
   console.error(error);
 
   return (
-    <Box>
+    <DefaultLayout>
       <Heading>Oops!</Heading>
       <Text>Sorry, an unexpected error has occurred.</Text>
       <Text>
         <Text as="i">{error.statusText || error.message}</Text>
       </Text>
-    </Box>
+    </DefaultLayout>
   );
 }
