+++ src/components/layout/index.ts
@@ -1,4 +1,3 @@
-export * from "./Footer";
-export * from "./Header";
-export * from "./LayoutFlex";
-export * from "./Main";
+/** @module Layout */
+export * from "./ErrorLayout";
+export * from "./DefaultLayout";
