+++ src/components/layout/navbar/index.ts
@@ -1,0 +1,2 @@
+export * from "./NavLink";
+export * from "./navbar.constants";
