+++ src/components/layout/partials/index.ts
@@ -1,2 +1,4 @@
-export * from "./NavLink";
-export * from "./navbar.constants";
+export * from "./Footer";
+export * from "./Header";
+export * from "./LayoutFlex";
+export * from "./Main";
